- Changed APlayerPawn::DamageFade to a PalEntry from 3 floats. - Removed #pragma warnings from cmdlib.h and fixed the places where they were   still triggered.   These #pragmas were responsible for >90% of the GCC warnings that were not   listed in VC++. - Fixed one bug in the process: DSeqNode::m_Atten was never adjusted when the   parameter handling of the sound functions for attenuation was changed.   Changed m_Atten to a float and fixed the SNDSEQ parser to set proper values.   Also added the option to specify attenuation with direct values in addition   to the predefined names.
SVN r1583 (trunk)

+++ src/c_cmds.cpp
@@ -623,7 +623,7 @@ CCMD (r_visibility)
 	}
 	else if (!netgame)
 	{
-		R_SetVisibility (atof (argv[1]));
+		R_SetVisibility ((float)atof (argv[1]));
 	}
 	else
 	{
